Implemented the ability to set data from mapped widget, when used manual submit.
This is a spike. The purpose of this spike is set data to the model when
focus leaves the widget. This functionality in such way must be implemented
for every widget, that is mapped with DB table field that is not foreign.

Also the app was named and there was corrected the "About app" text.

Signed-off-by: Oleg <olgerd27@gmail.com>